Add defaultTimerSeconds variable for hardcoded seconds
The timer for the game is now a variable instead of being hardcoded in the code.

Currently only the timer for a player's turn (=defaultTimerSeconds), for the fantastic four and for a counter attack.
All the others depend on defaultTimerSeconds, so if the players for example want to double the timer to 60 seconds, all the other timers are also doubled.

All the event timers are still hardcoded.